requirements.txt
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-oauthlib==1.3.0
pandas==2.3.3
psycopg2-binary==2.9.11
pydantic-settings==2.13.1
requests==2.33.0